airflow-pipeline/requirements.txt
apache-airflow==2.10.4
beautifulsoup4==4.13.3
pydantic==2.10.6
pandas==2.2.3
numpy==2.2.2
selenium==4.28.1
splinter==0.21.0
webdriver-manager==4.0.2
snowflake-connector-python
